@@ -1,9 +1,9 @@
 name: maybe_just_nothing
-version: 0.3.1
+version: 0.4.0-nullsafety.1
 homepage: https://github.com/f3ath/maybe-just-nothing
 description: Yet another variation of the good old Maybe monad with eager execution written in Dart. Or maybe it's just nothing.
 environment: 
-  sdk: '>=2.0.0 <3.0.0'
+  sdk: '>=2.10.0 <3.0.0'
 dev_dependencies: 
-  pedantic: ^1.9.0
-  test: ^1.9.2
+  pedantic: ^1.10.0-nullsafety.1
+  test: ^1.16.0-nullsafety.5
