@@ -5,7 +5,7 @@ import 'package:test/test.dart';
 
 void main() {
   test('Basic getters', () async {
-    Maybe<int> oddTimes3(int number) =>
+    Maybe<int> oddTimes3(int? number) =>
         Maybe(number).where((_) => _.isOdd).map((_) => _ * 3);
 
     expect(oddTimes3(5).orThrow(() => 'Oops'), 15);
@@ -21,13 +21,16 @@ void main() {
   test('Map', () {
     expect(Just(2).map((_) => _ * 2).orThrow(() => 'Oops'), 4);
     expect(Nothing<int>().map((_) => _ * 2), isA<Nothing<int>>());
-    expect(Just(2).map((_) => null), isA<Nothing<int>>());
   });
 
   test('Merge', () {
-    expect(Just(2).merge(Just(3), (a, b) => a * b).orThrow(() => 'Oops'), 6);
-    expect(Nothing<int>().merge(Just(3), (a, b) => a * b), isA<Nothing<num>>());
-    expect(Just(2).merge(Nothing<int>(), (a, b) => a * b), isA<Nothing<num>>());
+    expect(
+        Just(2).merge(Maybe(3), (int a, int b) => a * b).orThrow(() => 'Oops'),
+        6);
+    expect(Nothing<int>().merge(Just(3), (int a, int b) => a * b),
+        isA<Nothing<num>>());
+    expect(Just(2).merge(Nothing<int>(), (int a, int b) => a * b),
+        isA<Nothing<num>>());
   });
 
   test('FlatMap', () {
@@ -65,7 +68,7 @@ void main() {
   });
 
   test('Consumers', () {
-    int number;
+    int? number;
     void saveNumber(int n) => number = n;
     Nothing<int>().ifPresent(saveNumber);
     expect(number, isNull);
@@ -95,18 +98,22 @@ void main() {
     expect(Maybe(1) == Nothing<int>(), isFalse);
     // ignore: unrelated_type_equality_checks
     expect(Maybe(1) == Nothing(), isFalse);
+    // ignore: unrelated_type_equality_checks
     expect(Maybe(d) == Nothing(), isFalse);
     expect(Nothing<int>() == Maybe(1), isFalse);
     // ignore: unrelated_type_equality_checks
     expect(Nothing() == Maybe(1), isFalse);
     expect(Nothing<int>() == Nothing<int>(), isTrue);
+    // ignore: unrelated_type_equality_checks
+    expect(Nothing<int>() == Nothing<String>(), isFalse);
     expect(Nothing() == Nothing(), isTrue);
     expect(Nothing<int>() == Nothing(), isFalse);
     expect(Nothing() == Nothing<String>(), isFalse);
+    // ignore: unrelated_type_equality_checks
     expect(Nothing() == Maybe(d), isFalse);
 
-    int a;
-    String b;
+    int? a;
+    String? b;
     // ignore: unrelated_type_equality_checks
     expect(Maybe(a) == Maybe(b), isFalse);
   });
@@ -127,8 +134,8 @@ void main() {
     expect(Nothing().hashCode == Nothing<String>().hashCode, isFalse);
     expect(Nothing().hashCode == Maybe(d).hashCode, isFalse);
 
-    int a;
-    String b;
+    int? a;
+    String? b;
     expect(Maybe(a).hashCode == Maybe(b).hashCode, isFalse);
   });
 }
